Refactor for package privacy of emails
